pyproject.toml
[tool.poetry]
name = "fastapi_test_task"
version = "0.1.0"
description = ""
authors = ["Сергей <viip222@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.0"
SQLAlchemy = "^2.0.20"
alembic = "^1.11.3"
uvicorn = "^0.23.2"
psycopg2-binary = "^2.9.7"
python-dotenv = "^1.0.0"
asyncpg = "^0.28.0"
python-multipart = "^0.0.6"
fastapi-filter = "^0.6.1"
sqladmin = "^0.14.1"
itsdangerous = "^2.1.2"
pytest = "^7.4.0"
httpx = "^0.24.1"
pytest-asyncio = "^0.21.1"
isort = "^5.12.0"
gunicorn = "^21.2.0"

[tool.poetry.dev-dependencies]
flake8 = "^6.1.0"

[tool.poetry.group.dev.dependencies]
mimesis = "^11.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
